requirements.txt
peewee>=3.14.10
python-telegram-bot>=13.12
